pyproject.toml
[project]
name = "project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ollama>=0.6.0",
    "python-frontmatter>=1.1.0",
    "requests>=2.32.5",
    "weaviate-client>=4.17.0",
]

[dependency-groups]
dev = [
    "jupyter>=1.1.1",
]
